pyproject.toml
[project]
name = "aura"
version = "0.1.0"
description = "Sistema multiagentes de IA con RAG para recomendaciones de productos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.7.0",
    "sentence-transformers>=2.2.0",
]
